segmentVisualisationDashboard/requirements.txt
dash==2.15.0
dash-bootstrap-components==1.5.0
plotly==5.19.0
numpy==1.26.4
Pillow==10.2.0
opencv-python==4.9.0.80
gunicorn==21.2.0

